requirements.txt
absl-py==2.2.2
alembic==1.12.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
astunparse==1.6.3
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
billiard==4.2.1
black==23.11.0
cachetools==5.5.2
celery==5.3.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.104.1
flake8==6.1.0
flatbuffers==25.2.10
fonttools==4.58.0
gast==0.6.0
google-auth==2.40.2
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
greenlet==3.2.2
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
iniconfig==2.1.0
joblib==1.5.1
keras==2.15.0
kiwisolver==1.4.8
kombu==5.5.3
libclang==18.1.1
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.8.2
mccabe==0.7.0
ml-dtypes==0.2.0
# Editable install with no version control (music-analysis-system==0.1.0)
-e /home/anyone/.prjcts/music-analysis-system
mypy==1.7.1
mypy_extensions==1.1.0
numpy==1.25.2
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==25.0
pandas==2.1.3
passlib==1.7.4
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
plotly==5.17.0
pluggy==1.6.0
prompt_toolkit==3.0.51
protobuf==4.25.7
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.0.3
pydantic_core==2.14.1
pyflakes==3.1.0
pyparsing==3.2.3
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.15.3
seaborn==0.13.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==9.1.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.24.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.14.1
